Phase 0 audit remediation: crash fixes, dead-code purge, OSS compliance
Crash vectors fixed:
- Guard NSScreen.main ?? screens.first in AppDelegate launch path
- Safe fallbacks for global documents/temp dir + bundle ID constants
- NotificationWatcher: CFTypeID-check window attribute before AXUIElement cast
- AudioCaptureManager: degrade to flat bars if FFT setup fails (was fatalError)
- AudioPlayer: guard bundle resource URL and retain sound during playback

Dead code removed (~740 LOC): TestView/FluidSlider, DownloadView,
ProgressIndicator, BoringNotchWindow, BoringStatusMenu, EditPanelView,
sneakPeekEvent + SharedSneakPeek (contained as! Data crash), unused
BatteryActivityManager closures, dead dispatch items, 4 unused keyboard
shortcuts, dead sneak-peek size constants, NotchSpaceManager leftovers.
VisualEffectView extracted into its own file (was hiding in deleted
EditPanelView.swift, used by 5 onboarding views).

OSS compliance & hygiene:
- THIRD_PARTY_LICENSES: add all 12 SPM packages (Apache-2.0 block for
  lottie-spm/swift-collections/swift-syntax, MIT block +7, BSD block +1)
- Fix broken THIRD_PARTY_LICENSES link in README
- mediaremote-adapter: add provenance README (pinned at upstream v0.7.2,
  verified byte-identical script, rebuild/update instructions)
- Rename PR template so GitHub actually renders it; delete 2 stale
  issue templates; fix CONTRIBUTING step numbering; drop Linux devcontainer
- Commit shared boringNotch scheme; clear hardcoded DEVELOPMENT_TEAM
  (CI overrides it); clean up .gitignore (stop ignoring schemes)

Verified: xcodebuild Debug build succeeds for app + XPC helper.
