--- .devcontainer/devcontainer.json
@@ -1,5 +1,0 @@
-{
-  "image": "mcr.microsoft.com/devcontainers/universal:2",
-  "features": {
-  }
-}
