--- .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,21 +1,0 @@
-<!--- This issue template is being replaced with the updated issue form for new features -->
----
-name: Feature request
-about: Suggest an idea for this project
-title: "[FEATURE]"
-labels: ''
-assignees: ''
-
----
-
-**Is your feature request related to a problem? Please describe.**
-A clear and concise description of what the problem is. Ex. I'm always frustrated when [...]
-
-**Describe the solution you'd like**
-A clear and concise description of what you want to happen.
-
-**Additional context**
-Add any other context or screenshots about the feature request here.
-
-**Checks**
-- [x] I haven't found any duplicates with my issue.
