--- .github/ISSUE_TEMPLATE/old_bug_report.md
@@ -1,31 +1,0 @@
-<!--- This issue template is deprecated, use the updated issue form for bug reports -->
----
-name: Bug report
-about: Create a report to help us improve
-title: "[BUG]"
-labels: 'bug,unconfirmed'
-assignees: ''
-
----
-
-**Describe the bug**
-A clear and concise description of what the bug is.
-
-**To Reproduce**
-Steps to reproduce the behavior:
-1. Go to '...'
-2. Click on '....'
-3. Scroll down to '....'
-4. See error
-
-**Expected behavior**
-A clear and concise description of what you expected to happen.
-
-**Screenshots or recordings**
-If applicable, add screenshots to help explain your problem.
-
-**Additional context**
-Add any other context about the problem here. Mentioning what version are you using.
-
-**Checks**
-- [x] I haven't found any duplicates with my issue.
