+++ boringNotch/Shortcuts/ShortcutConstants.swift
@@ -1,5 +1,5 @@
 //
-//  Constants.swift
+//  ShortcutConstants.swift
 //  boringNotch
 //
 //  Created by Richard Kunkli on 16/08/2024.
@@ -9,10 +9,6 @@ import KeyboardShortcuts
 import SwiftUI
 
 extension KeyboardShortcuts.Name {
-    static let clipboardHistoryPanel = Self("clipboardHistoryPanel", default: .init(.c, modifiers: [.shift, .command]))
-    static let toggleMicrophone = Self("toggleMicrophone", default: .init(.f5, modifiers: [.function]))
-    static let decreaseBacklight = Self("decreaseBacklight", default: .init(.f1, modifiers: [.command]))
-    static let increaseBacklight = Self("increaseBacklight", default: .init(.f2, modifiers: [.command]))
     static let toggleSneakPeek = Self("toggleSneakPeek", default: .init(.h, modifiers: [.command, .shift]))
     static let toggleNotchOpen = Self("toggleNotchOpen", default: .init(.i, modifiers: [.command, .shift]))
 }
