+++ boringNotch/managers/NotchSpaceManager.swift
@@ -10,9 +10,7 @@ import Foundation
 class NotchSpaceManager {
     static let shared = NotchSpaceManager()
     let notchSpace: CGSSpace
-    private var eventTap: CFMachPort?
-    private var runLoopSource: CFRunLoopSource?
-    
+
     private init() {
         notchSpace = CGSSpace(level: 2147483647) // Max level
     }
