+++ boringNotch/sizing/matters.swift
@@ -9,9 +9,6 @@ import Defaults
 import Foundation
 import SwiftUI
 
-let downloadSneakSize: CGSize = .init(width: 65, height: 1)
-let batterySneakSize: CGSize = .init(width: 160, height: 1)
-
 let shadowPadding: CGFloat = 20
 let openNotchSize: CGSize = .init(width: 640, height: 190)
 let windowSize: CGSize = .init(width: openNotchSize.width, height: openNotchSize.height + shadowPadding)
